[ci skip] Add `write` methods to CDS and fix MRT writer for empty metadata
Byte-by-byte is exactly the same in CDS and MRT, but metadata is not
